move components to lending global components
