+++ src/pages/lending/hooks/table/use-table-my-supplies.tsx
@@ -26,7 +26,7 @@ import { useTableLendingMySuppliesSortStore } from '~/states/components';
 import { POPUP_ID, TOOLTIP_ID } from '~/types';
 
 import { mySuppliesData } from '../../data';
-import { APYSmall } from '../../pages/main/components/apy';
+import { APYSmall } from '../../components/apy';
 
 export const useTableMySupplies = () => {
   const { sort, setSort } = useTableLendingMySuppliesSortStore();
